move random_choice_csc to utils/random from utils/sparsefuncs
